+++ CHANGELOG.md
@@ -25,5 +25,6 @@ Initial release.
 
 - Support Agda 2.6.3; remove support for earlier versions.
 - Support irrefutable let-patterns (fixes bug).
+- Handle variable shadowing correctly in let-patterns (fixes bug).
 - Resolve imports correctly within the current project (fixes bug).